Move conversation request up one level
