+++ src/components/Messages/ConversationList.tsx
@@ -1,10 +1,8 @@
 import ChatOutlinedIcon from "@mui/icons-material/ChatOutlined";
 import { Box, Divider, IconButton, List, Typography } from "@mui/material";
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import { useNavigate } from "react-router-dom";
-import { useAppDispatch, useAppSelector } from "../../state/hooks";
-import { setConversations } from "../../state/slices/messagesSlice";
-import useAxios from "../../utilities/useAxios";
+import { useAppSelector } from "../../state/hooks";
 import SearchBar from "../Common/SearchBar";
 import ConversationListItem from "./ConversationListItem";
 import NewMessageModal from "./NewMessageModal/NewMessageModal";
@@ -29,24 +27,8 @@ const ConversationList = () => {
     (state) => state.messages.selectedConversation.userId,
   );
   const userId = useAppSelector((state) => state.user.userId);
-  const dispatch = useAppDispatch();
 
   const navigate = useNavigate();
-  const { sendRequest } = useAxios();
-
-  useEffect(() => {
-    const fetchMessages = async () => {
-      const result = await sendRequest(
-        {
-          method: "GET",
-          params: { userId },
-        },
-        "messages",
-      );
-      dispatch(setConversations(result));
-    };
-    fetchMessages();
-  }, [dispatch, userId, sendRequest]);
 
   return (
     <>
